+++ src/pages/login/email-register.tsx
@@ -6,11 +6,11 @@ import {
   useUserLoginErrorStore,
   useUserLoginStore,
 } from '@zustand/usersLoginStore';
-import useCheckPasswords from 'hooks/useCheckPasswords';
-import useEmailSignUpValidation from 'hooks/useEmailSignUpValidation';
-import useGetSessionUserInfo from 'hooks/useGetSessionUserInfo';
-import useHandleLoginUser from 'hooks/useHandleLoginUser';
-import useInitLoginUser from 'hooks/useInitLoginUser';
+import useCheckPasswords from 'hooks/login/useCheckPasswords';
+import useEmailSignUpValidation from 'hooks/login/useEmailSignUpValidation';
+import useHandleLoginUser from 'hooks/login/useHandleLoginUser';
+import useInitLoginUser from 'hooks/login/useInitLoginUser';
+
 import { SESSION_KEY, setSessionStorage } from 'libraries/sessionStorageUtils';
 import { useRouter } from 'next/router';
 import { signUpEmail } from 'pages/api/login/auth';
