+++ src/pages/login/social-email-register.tsx
@@ -6,9 +6,9 @@ import {
   useUserLoginErrorStore,
   useUserLoginStore,
 } from '@zustand/usersLoginStore';
-import useGetSessionUserInfo from 'hooks/useGetSessionUserInfo';
-import useHandleLoginUser from 'hooks/useHandleLoginUser';
-import useInitLoginUser from 'hooks/useInitLoginUser';
+import useGetSessionUserInfo from 'hooks/login/useGetSessionUserInfo';
+import useHandleLoginUser from 'hooks/login/useHandleLoginUser';
+
 import { useRouter } from 'next/router';
 
 export default function SocialEmailRegister() {
