+++ src/pages/login/twitter-email-register.tsx
@@ -6,9 +6,10 @@ import {
   useUserLoginErrorStore,
   useUserLoginStore,
 } from '@zustand/usersLoginStore';
-import useCheckPasswords from 'hooks/useCheckPasswords';
-import useHandleLoginUser from 'hooks/useHandleLoginUser';
-import useInitLoginUser from 'hooks/useInitLoginUser';
+import useCheckPasswords from 'hooks/login/useCheckPasswords';
+import useHandleLoginUser from 'hooks/login/useHandleLoginUser';
+import useInitLoginUser from 'hooks/login/useInitLoginUser';
+
 import { extractEmailPart } from 'libraries/stringUtils';
 
 import { useRouter } from 'next/router';
@@ -21,7 +22,6 @@ export default function TwitterEmailRegister() {
   const { handleChangeEmail } = useHandleLoginUser();
 
   const handleClickNext = () => {
-    //if(provider === 'twitter') {
     setLoginState({ name: extractEmailPart(email) });
     router.push('/login/nickname');
   };
